Code refactoring: expand chart to new class by their type
